+++ rabbitmq/clowder_list/rabbitmq_handler.py
@@ -6,6 +6,8 @@ import pika
 import requests
 
 RABBITMQ_HOST = os.getenv('RABBITMQ_HOST', 'rabbitmq')
+RABBITMQ_USER = os.getenv('RABBITMQ_HOST', 'guest')
+RABBITMQ_PASSWORD = os.getenv('RABBITMQ_HOST', 'guest')
 
 
 def rabbitmq_handler(ch, method, properties, body):
@@ -54,7 +56,9 @@ def rabbitmq_handler(ch, method, properties, body):
 
 
 if __name__ == '__main__':
-    connection = pika.BlockingConnection(pika.ConnectionParameters(port=5672, host=RABBITMQ_HOST))
+    credentials = pika.PlainCredentials(RABBITMQ_USER, RABBITMQ_PASSWORD)
+    parameters = pika.ConnectionParameters(RABBITMQ_HOST, 5672, '/', credentials)
+    connection = pika.BlockingConnection(parameters)
     channel = connection.channel()
 
     # pass the queue name in environment variable