+++ rabbitmq/name_entity_recognition/rabbitmq_handler.py
@@ -5,6 +5,9 @@ import pika
 import postToAWSBatch
 
 RABBITMQ_HOST = os.getenv('RABBITMQ_HOST', 'rabbitmq')
+RABBITMQ_USER = os.getenv('RABBITMQ_USER', 'guest')
+RABBITMQ_PASSWORD = os.getenv('RABBITMQ_PASSWORD', 'guest')
+
 
 def rabbitmq_handler(ch, method, properties, body):
     try:
@@ -51,11 +54,13 @@ def rabbitmq_handler(ch, method, properties, body):
 
 
 if __name__ == '__main__':
+    credentials = pika.PlainCredentials(RABBITMQ_USER, RABBITMQ_PASSWORD)
     connection = pika.BlockingConnection(pika.ConnectionParameters(
         port=5672,
         host=RABBITMQ_HOST,
         heartbeat=600,
-        blocked_connection_timeout=600
+        blocked_connection_timeout=600,
+        credentials=credentials
     ))
     channel = connection.channel()