+++ rabbitmq/topic_modeling/rabbitmq_handler.py
@@ -5,6 +5,9 @@ import pika
 import postToAWSBatch
 
 RABBITMQ_HOST = os.getenv('RABBITMQ_HOST', 'rabbitmq')
+RABBITMQ_USER = os.getenv('RABBITMQ_USER', 'guest')
+RABBITMQ_PASSWORD = os.getenv('RABBITMQ_PASSWORD', 'guest')
+
 
 def rabbitmq_handler(ch, method, properties, body):
     try:
@@ -52,7 +55,9 @@ def rabbitmq_handler(ch, method, properties, body):
 
 
 if __name__ == '__main__':
-    connection = pika.BlockingConnection(pika.ConnectionParameters(port=5672, host=RABBITMQ_HOST))
+    credentials = pika.PlainCredentials(RABBITMQ_USER, RABBITMQ_PASSWORD)
+    parameters = pika.ConnectionParameters(RABBITMQ_HOST, 5672, '/', credentials)
+    connection = pika.BlockingConnection(parameters)
     channel = connection.channel()
 
     # pass the queue name in environment variable