+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Docker building script for whole components [23](https://github.com/ncsa/standalone-smm-analytics/issues/23)
-
+- Docker compose launch script [45](https://github.com/ncsa/standalone-smm-analytics/issues/45)
+- Docker compose file using traefik [46](https://github.com/ncsa/standalone-smm-analytics/issues/46)
+- 
 ### Changed
 - Hard coded rabbimq url changed to env variable [18](https://github.com/ncsa/standalone-smm-analytics/issues/18)
-- Modifed S3 url to env variable [21](https://github.com/ncsa/standalone-smm-analytics/issues/21)
+- Modified S3 url to env variable [21](https://github.com/ncsa/standalone-smm-analytics/issues/21)
 - Renamed Minio related environment variables [31](https://github.com/ncsa/standalone-smm-analytics/issues/31)
+- Rabbitmq handler's connection with dynamic credentials [41](https://github.com/ncsa/standalone-smm-analytics/issues/41)
+- Docker compose file to work with new settings [42](https://github.com/ncsa/standalone-smm-analytics/issues/42)